requirements.txt
# Základní závislosti
python-dotenv>=1.0.0
requests>=2.31.0

# Notion API
notion-client>=2.0.0

# NLP knihovny
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.35.0
torch>=2.0.0

# Pomocné knihovny
pydantic>=2.4.0
tqdm>=4.66.0

# Testování
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting a formátování
black>=23.9.0
pylint>=3.0.0
isort>=5.12.0
